Move CountRanges*() to a separate object file
It needs to be usable in Linux subagent shared library, but also in the
unit test executable.
