+++ src/agent/subagents/linux/Makefile.am
@@ -1,7 +1,7 @@
 SUBAGENT = linux
 
 pkglib_LTLIBRARIES = linux.la
-linux_la_SOURCES = cpu.cpp disk.cpp drbd.cpp hardware.cpp hypervisor.cpp \
+linux_la_SOURCES =  count_ranges.cpp cpu.cpp disk.cpp drbd.cpp hardware.cpp hypervisor.cpp \
                    iostat.cpp linux.cpp net.cpp packages.cpp proc.cpp system.cpp
 linux_la_CPPFLAGS=-I@top_srcdir@/include -I@top_srcdir@/build
 linux_la_LDFLAGS = -module -avoid-version -export-symbols ../platform-subagent.sym
