+++ src/agent/subagents/linux/cpu.h
@@ -314,71 +314,3 @@ float Collector::getCoreUsage(enum CpuUsageSource source, int coreIndex, int nbL
 
    return core.m_tables[source].getAverage(nbLastItems);
 }
-
-/**
- * Count items in a list of ranges, like e.g. /sys/devices/system/cpu/online
- *
- * @param buffer contents of the file verbatim. The buffer is destructively modified!
- * @return count of items (e.g. CPU) as described in the buffer
- */
-long CountRanges(char *buffer)
-{
-   char *endline = strchrnul(buffer, '\n');
-   *endline = 0;
-
-   int count = 0;
-   char *rangeStartPtr = buffer;
-   while (true) {
-      char *rangeEndPtr = strchrnul(rangeStartPtr, ',');
-      bool last = *rangeEndPtr == '\0'; // true if this is the last or the only range
-      *rangeEndPtr = 0;
-      // Now rangeStartPtr points at null-terminated string with a single range of non-negative numbers
-      char *afterStartIndex;
-      long startIndex = strtol(rangeStartPtr, &afterStartIndex, 10);
-      if (afterStartIndex == rangeStartPtr)
-      {
-         ; // empty range
-      }
-      else if (*afterStartIndex == '\0')
-      {
-         assert(startIndex >= 0);
-         count++;
-      }
-      else if (*afterStartIndex == '-')
-      {
-         long endIndex = strtol(afterStartIndex + 1 /* skip minus sign */, &afterStartIndex, 10);
-         assert(endIndex >= 0);
-         assert(*afterStartIndex == '\0');
-         count += endIndex - startIndex + 1;
-      }
-      else
-      {
-         assert(false);
-      }
-
-      if (last)
-         break;
-      rangeStartPtr = rangeEndPtr + 1;
-   }
-   return count;
-}
-
-/**
- * Count items in a ranges-list file, like e.g. /sys/devices/system/cpu/online
- */
-static long CountRangesFile(const char *filepath)
-{
-   FILE *f = fopen(filepath, "r");
-   if (f == nullptr)
-   {
-      nxlog_write_tag(NXLOG_WARNING, DEBUG_TAG, _T("Cannot open file %s"), filepath);
-      return 0;
-   }
-
-   char buffer[256];
-   if (fgets(buffer, sizeof(buffer), f) == nullptr)
-      return 0;
-   fclose(f);
-
-   return CountRanges(buffer);
-}
