+++ tests/test-unit-linux-cpu-usage-collector/Makefile.am
@@ -9,7 +9,7 @@
 # implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
 
 bin_PROGRAMS = test-unit-linux-cpu-usage-collector
-test_unit_linux_cpu_usage_collector_SOURCES = cpu.cpp main.cpp
+test_unit_linux_cpu_usage_collector_SOURCES = cpu.cpp main.cpp ../../src/agent/subagents/linux/count_ranges.cpp
 test_unit_linux_cpu_usage_collector_CPPFLAGS = -I@top_srcdir@/include -I../include -I@top_srcdir@/build
 test_unit_linux_cpu_usage_collector_LDFLAGS = @EXEC_LDFLAGS@
 test_unit_linux_cpu_usage_collector_LDADD = @top_srcdir@/src/libnetxms/libnetxms.la @EXEC_LIBS@
